Cache XML elementrees.
doxy.DoxyDB.get_tree() was a major contibutor to the execution time of
the doxy:c directive. Reading and parsing the XML was done for each entity.

In practice, there is usually a succession of directive that document entities
living in the same file.

By using functools.lru_cache it is possible to reduce the build times of
examples/RIOT from 210 seconds to 174.