Manually re-root enumvalues under enums.
For some wird reason, enumvalues are defined inside enums, but in the
index they show up as children of files. I suspect this is part of the
reason for the Doxygen bug that causes it to repeat enum values and
mess them up.

For now we can manually change the hierarchy. That will allow us to
manually fix repeated definitions using sphinx events.