XSLT: work around doxygen bugs.
Doxygen does not generate <param> elements for arguments to function
pointers. We work around it by adding a custom XSLT function that creates
desc_parameter/_type/_name nodes from the text signature.

This is not perfect (it will not have hyperlinks) but it is better than
nothing.

For more info see these bugs:

https://github.com/doxygen/doxygen/issues/4668
https://github.com/doxygen/doxygen/issues/5542